requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
h11>=0.14.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic-settings>=2.2.0
elasticsearch[async]>=8.12.0,<9
python-dotenv>=1.0.0
